// Fitlytics/src/services/generated/models/WorkoutProgramDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProgramWeekDto } from './ProgramWeekDto';
import {
    ProgramWeekDtoFromJSON,
    ProgramWeekDtoFromJSONTyped,
    ProgramWeekDtoToJSON,
    ProgramWeekDtoToJSONTyped,
} from './ProgramWeekDto';

/**
 * 
 * @export
 * @interface WorkoutProgramDto
 */
export interface WorkoutProgramDto {
    /**
     * 
     * @type {string}
     * @memberof WorkoutProgramDto
     */
    workoutProgramId?: string;
    /**
     * 
     * @type {string}
     * @memberof WorkoutProgramDto
     */
    name: string | null;
    /**
     * 
     * @type {Array<ProgramWeekDto>}
     * @memberof WorkoutProgramDto
     */
    programWeeks: Array<ProgramWeekDto> | null;
}

/**
 * Check if a given object implements the WorkoutProgramDto interface.
 */
export function instanceOfWorkoutProgramDto(value: object): value is WorkoutProgramDto {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('programWeeks' in value) || value['programWeeks'] === undefined) return false;
    return true;
}

export function WorkoutProgramDtoFromJSON(json: any): WorkoutProgramDto {
    return WorkoutProgramDtoFromJSONTyped(json, false);
}

export function WorkoutProgramDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkoutProgramDto {
    if (json == null) {
        return json;
    }
    return {
        
        'workoutProgramId': json['workoutProgramId'] == null ? undefined : json['workoutProgramId'],
        'name': json['name'],
        'programWeeks': (json['programWeeks'] == null ? null : (json['programWeeks'] as Array<any>).map(ProgramWeekDtoFromJSON)),
    };
}

export function WorkoutProgramDtoToJSON(json: any): WorkoutProgramDto {
    return WorkoutProgramDtoToJSONTyped(json, false);
}

export function WorkoutProgramDtoToJSONTyped(value?: WorkoutProgramDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'workoutProgramId': value['workoutProgramId'],
        'name': value['name'],
        'programWeeks': (value['programWeeks'] == null ? null : (value['programWeeks'] as Array<any>).map(ProgramWeekDtoToJSON)),
    };
}