// Fitlytics/src/services/generated/models/WorkoutDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WorkoutExerciseDto } from './WorkoutExerciseDto';
import {
    WorkoutExerciseDtoFromJSON,
    WorkoutExerciseDtoFromJSONTyped,
    WorkoutExerciseDtoToJSON,
    WorkoutExerciseDtoToJSONTyped,
} from './WorkoutExerciseDto';

/**
 * 
 * @export
 * @interface WorkoutDto
 */
export interface WorkoutDto {
    /**
     * 
     * @type {string}
     * @memberof WorkoutDto
     */
    workoutId?: string;
    /**
     * 
     * @type {string}
     * @memberof WorkoutDto
     */
    name: string | null;
    /**
     * 
     * @type {Array<WorkoutExerciseDto>}
     * @memberof WorkoutDto
     */
    workoutExercises: Array<WorkoutExerciseDto> | null;
}

/**
 * Check if a given object implements the WorkoutDto interface.
 */
export function instanceOfWorkoutDto(value: object): value is WorkoutDto {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('workoutExercises' in value) || value['workoutExercises'] === undefined) return false;
    return true;
}

export function WorkoutDtoFromJSON(json: any): WorkoutDto {
    return WorkoutDtoFromJSONTyped(json, false);
}

export function WorkoutDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkoutDto {
    if (json == null) {
        return json;
    }
    return {
        
        'workoutId': json['workoutId'] == null ? undefined : json['workoutId'],
        'name': json['name'],
        'workoutExercises': (json['workoutExercises'] == null ? null : (json['workoutExercises'] as Array<any>).map(WorkoutExerciseDtoFromJSON)),
    };
}

export function WorkoutDtoToJSON(json: any): WorkoutDto {
    return WorkoutDtoToJSONTyped(json, false);
}

export function WorkoutDtoToJSONTyped(value?: WorkoutDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'workoutId': value['workoutId'],
        'name': value['name'],
        'workoutExercises': (value['workoutExercises'] == null ? null : (value['workoutExercises'] as Array<any>).map(WorkoutExerciseDtoToJSON)),
    };
}